Rename wifi connect routes files
